Fix bug related to colors, use free()
- Fixed bug related to migration to fixed-sized types concerning colors.
- Use free() on data structures after we're done using them.

+++ colors.h
@@ -36,7 +36,7 @@ struct RGB {
 };
 
 struct Theme {
-	uint16_t colors[MAXCOLORS];
+	uint32_t colors[MAXCOLORS];
 	uint16_t numColors;
 	uint8_t  nameLen;
 	char*    name;
